+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#101] Generation of wait steps for dogus which have optional component dependencies.
 
 ## [v2.1.0] - 2024-09-18
 ### Changed